requirements_algo.txt
# March 24: use py <= 3.9.12
# to resolve C++ build errors:
# https://stackoverflow.com/questions/64261546/how-to-solve-error-microsoft-visual-c-14-0-or-greater-is-required-when-inst
# The solution - manually install these packages before installing alpaca-trade-api. e.g:
pip install pandas==1.1.5 numpy==1.19.4 scipy==1.5.4
# DEPENDENCIES USED IN ALL ENVIRONMENTS (without ML for ma_loop only)
# do not use tabs here when copied;
Backtesting==0.3.3
# needed for py 3.10.0
numpy==1.21.6
pandas==2.0.0
scipy==1.10.1
Keras-Preprocessing==1.1.2
Markdown==3.3.4
Pillow==8.2.0
Werkzeug==2.0.1
absl-py==0.12.0
aiohttp==3.8.1
alpaca-py==0.30.1
alpaca-trade-api==2.3.0
alpha-vantage==2.3.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==21.2.0
beautifulsoup4==4.9.1
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
chardet==3.0.4
cycler==0.10.0
gast==0.4.0
google-auth==1.30.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.38.0
h5py==3.2.1
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.1
msgpack==1.0.3
oauthlib==3.1.1
opt-einsum==3.3.0
protobuf==3.17.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
setuptools==57.0.0
six==1.16.0
soupsieve==2.2.1
termcolor==1.1.0
threadpoolctl==2.1.0
urllib3==1.26.5
websocket-client==1.0.1
websockets==9.1
wheel==0.36.2
wrapt==1.12.1